+++ app/Apis/Summ_Ai/Request.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * File for handler for each request to SUMM AI API.
+ * File for a handler for each request to SUMM AI API.
  *
  * @doc https://backend.summ-ai.com/api/docs/
  *
@@ -17,6 +17,7 @@ use easyLanguage\Plugin\Log;
 use easyLanguage\Plugin\Log_Api;
 use easyLanguage\EasyLanguage\Db;
 use WP_Error;
+use WP_HTTP_Requests_Response;
 
 /**
  * Create and send a request to SUMM AI API.
@@ -307,7 +308,10 @@ class Request implements Api_Requests {
 			$this->response = wp_remote_retrieve_body( $this->get_result() );
 
 			// secure http-status.
-			$this->http_status = $this->get_result()['http_response']->get_status(); // @phpstan-ignore offsetAccess.nonOffsetAccessible
+			$result = $this->get_result();
+			if ( is_array( $result ) && ! empty( $result['http_response'] ) && $result['http_response'] instanceof WP_HTTP_Requests_Response ) {
+				$this->http_status = $result['http_response']->get_status();
+			}
 
 			// log the request (with an anonymized token).
 			$args['headers']['Authorization'] = 'anonymized';
@@ -317,7 +321,7 @@ class Request implements Api_Requests {
 			}
 			Log_Api::get_instance()->add_log( $summ_ai_obj->get_name(), $this->http_status, $args_json, 'HTTP-Status: ' . $this->get_http_status() . '<br>' . $this->response );
 
-			// save the request and result in db.
+			// save the request and result in the database.
 			$this->save_in_db();
 		}
 	}
@@ -363,7 +367,7 @@ class Request implements Api_Requests {
 	/**
 	 * Get the complete request results.
 	 *
-	 * @return array<string>|WP_Error
+	 * @return array<string,mixed>|WP_Error
 	 */
 	public function get_result(): WP_Error|array {
 		return $this->result;
@@ -436,7 +440,7 @@ class Request implements Api_Requests {
 	/**
 	 * Set source-language for this request.
 	 *
-	 * @param string $language The source language as string (e.g. "de_EL").
+	 * @param string $language The source language as string (e.g., "de_EL").
 	 *
 	 * @return void
 	 * @noinspection PhpUnused
@@ -446,7 +450,7 @@ class Request implements Api_Requests {
 	}
 
 	/**
-	 * Save the data of this request in DB.
+	 * Save the data of this request in the database.
 	 *
 	 * @return void
 	 */