+++ app/Plugin/Apis.php
@@ -218,6 +218,13 @@ class Apis {
 		// check nonce.
 		check_admin_referer( 'easy-language-clear-api-log', 'nonce' );
 
+		// bail if user has not the capability for this.
+		if ( ! current_user_can( Settings::get_instance()->get_settings_obj()->get_capability() ) ) {
+			// redirect user back.
+			wp_safe_redirect( wp_get_referer() );
+			exit;
+		}
+
 		// get name of the api to clear.
 		$clear_api = isset( $_GET['api'] ) ? sanitize_text_field( wp_unslash( $_GET['api'] ) ) : '';